# openapi.yaml
openapi: 3.0.3
info:
  title: The Vanity E-commerce API
  description: |
    RESTful API for The Vanity E-commerce Platform.
    This API provides endpoints for managing products, users, orders, and admin operations.
  version: 1.0.0
  contact:
    name: The Vanity Support
    email: support@thevanity.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api
    description: Local development server
  - url: https://api.thevanity.com/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
  schemas:
    Product:
      type: object
      required:
        - name
        - brand
        - price
        - image
        - category
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        name:
          type: string
          example: "Crystal Pendant Necklace Set"
        brand:
          type: string
          example: "Fashion Jewelry"
        price:
          type: number
          example: 600
        originalPrice:
          type: number
          example: 899
        image:
          type: string
          format: uri
          example: "https://example.com/image.jpg"
        images:
          type: array
          items:
            type: string
            format: uri
        category:
          type: string
          example: "Necklaces"
        description:
          type: string
          example: "Elegant crystal pendant necklace set with intricate detailing"
        reviews:
          type: integer
          default: 0
          example: 120
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          default: 0
          example: 4.5
        stock:
          type: integer
          default: 0
          example: 45
        specifications:
          type: object
          additionalProperties: true
        isPublic:
          type: boolean
          default: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    User:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439011"
        firstName:
          type: string
          example: "Priya"
        lastName:
          type: string
          example: "Sharma"
        email:
          type: string
          format: email
          example: "priya.sharma@email.com"
        phone:
          type: string
          example: "+91 98765 43210"
        role:
          type: string
          enum: [user, admin]
          default: user
        status:
          type: string
          enum: [active, suspended]
          default: active
        emailVerified:
          type: boolean
          default: false
        phoneVerified:
          type: boolean
          default: false
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/Address'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
...
    Order:
      type: object
      required:
        - orderNumber
        - userId
        - items
        - totalAmount
      properties:
        _id:
          type: string
        orderNumber:
          type: string
          example: "ORD-2024-001"
        userId:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              name:
                type: string
              quantity:
                type: integer
              price:
                type: number
        totalAmount:
          type: number
          example: 1200
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
          default: pending
        shippingAddress:
          $ref: '#/components/schemas/Address'
        paymentMethod:
          type: string
          enum: [cod, card, upi, netbanking]
        paymentStatus:
          type: string
          enum: [pending, completed, failed, refunded]
        trackingNumber:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Category:
      type: object
      required:
        - name
      properties:
        _id:
          type: string
        name:
          type: string
          example: "Makeup"
        slug:
          type: string
          example: "makeup"
        description:
          type: string
        parentId:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true
        sortOrder:
          type: integer
          default: 0
        image:
          type: string
          format: uri
        attributes:
          type: array
          items:
            type: string
        productCount:
          type: integer
          readOnly: true
        children:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Inventory:
      type: object
      properties:
        _id:
          type: string
        productId:
          type: string
        productName:
          type: string
        sku:
          type: string
        category:
          type: string
        currentStock:
          type: integer
        reservedStock:
          type: integer
        availableStock:
          type: integer
        reorderLevel:
          type: integer
        maxStock:
          type: integer
        unitCost:
          type: number
        totalValue:
          type: number
        status:
          type: string
          enum: [in_stock, low_stock, out_of_stock, discontinued]
        supplier:
          type: string
        location:
          type: string
        movements:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              type:
                type: string
                enum: [sale, restock, adjustment, return]
              quantity:
                type: integer
              reason:
                type: string
        lastUpdated:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Error message"
        statusCode:
          type: integer
          example: 400
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve a list of all products with optional filtering
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category name
        - name: limit
          in: query
          schema:
            type: integer
          description: Limit number of results
        - name: exclude
          in: query
          schema:
            type: string
          description: Exclude product by ID
        - name: search
          in: query
          schema:
            type: string
          description: Search in product name, brand, category
        - name: minPrice
          in: query
          schema:
            type: number
          description: Minimum price filter
        - name: maxPrice
          in: query
          schema:
            type: number
          description: Maximum price filter
        - name: brand
          in: query
          schema:
            type: string
          description: Filter by brand
        - name: isPublic
          in: query
          schema:
            type: boolean
            default: true
          description: Filter by public/private status
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Get single product
      description: Retrieve a single product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Product ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== CATEGORIES ====================
...
  /admin/users:
    get:
      tags:
        - Admin - Users
      summary: Get all users (Admin)
      description: Retrieve all users with optional filtering
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
...
  /admin/inventory:
    get:
      tags:
        - Admin - Inventory
      summary: Get inventory
      description: Retrieve all inventory items (admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Inventory'
  /admin/inventory/{id}:
    put:
      tags:
        - Admin - Inventory
      summary: Update inventory
      description: Update inventory item (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentStock:
                  type: integer
                reorderLevel:
                  type: integer
      responses:
        '200':
          description: Inventory updated successfully

  # ==================== ADMIN - CATEGORIES ====================
  /admin/categories:
    get:
      tags:
        - Admin - Categories
      summary: Get all categories (Admin)
      description: Retrieve all categories with admin details
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

    post:
      tags:
        - Admin - Categories
      summary: Create category
      description: Create a new category (admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '201':
          description: Category created successfully
  /admin/categories/{id}:
    put:
      tags:
        - Admin - Categories
      summary: Update category
      description: Update existing category (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Category updated successfully

    delete:
      tags:
        - Admin - Categories
      summary: Delete category
      description: Delete a category (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Category deleted successfully

  # ==================== ADMIN - ORDERS ====================
  /admin/orders:
    get:
      tags:
        - Admin - Orders
      summary: Get all orders (Admin)
      description: Retrieve all orders with optional filtering
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /admin/orders/{id}/status:
    put:
      tags:
        - Admin - Orders
      summary: Update order status
      description: Update order status and tracking (admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, processing, shipped, delivered, cancelled]
                trackingNumber:
                  type: string
      responses:
        '200':
          description: Order status updated successfully

  # ==================== ADMIN - NOTIFICATIONS ====================
  /admin/notifications:
    get:
      tags:
        - Admin - Notifications
      summary: Get notifications
      description: Retrieve admin notifications
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
  /admin/notifications/{id}/read:
    put:
      tags:
        - Admin - Notifications
      summary: Mark notification as read
      description: Mark a notification as read
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read